# Avoid thread-local state when computing CRCs