--- a/bookkeeper-server/src/main/java/org/apache/bookkeeper/proto/checksum/CRC32CDigestManager.java
+++ b/bookkeeper-server/src/main/java/org/apache/bookkeeper/proto/checksum/CRC32CDigestManager.java
@@ -28,13 +28,6 @@ import org.apache.commons.lang3.mutable.MutableInt;
 @Slf4j
 class CRC32CDigestManager extends DigestManager {
 
-    private static final FastThreadLocal<MutableInt> currentCrc = new FastThreadLocal<MutableInt>() {
-        @Override
-        protected MutableInt initialValue() throws Exception {
-            return new MutableInt(0);
-        }
-    };
-
     public CRC32CDigestManager(long ledgerId, boolean useV2Protocol, ByteBufAllocator allocator) {
         super(ledgerId, useV2Protocol, allocator);
     }
@@ -45,16 +38,17 @@ class CRC32CDigestManager extends DigestManager {
     }
 
     @Override
-    void populateValueAndReset(ByteBuf buf) {
-        MutableInt current = currentCrc.get();
-        buf.writeInt(current.intValue());
-        current.setValue(0);
+    boolean isInt32Digest() {
+        return true;
     }
 
     @Override
-    void update(ByteBuf data, int offset, int len) {
-        MutableInt current = currentCrc.get();
-        final int lastCrc = current.intValue();
-        current.setValue(Crc32cIntChecksum.resumeChecksum(lastCrc, data, offset, len));
+    void populateValueAndReset(int digest, ByteBuf buf) {
+        buf.writeInt(digest);
+    }
+
+    @Override
+    int update(int digest, ByteBuf data, int offset, int len) {
+        return Crc32cIntChecksum.resumeChecksum(digest, data, offset, len);
     }
 }